+++ react/components/SliderTrack.tsx
@@ -155,7 +155,10 @@ const SliderTrack: FC<Props> = ({
           })
           groupDispatch?.({
             type: 'SLIDE',
-            payload: { currentSlide: 0, transform: transformMap[0] },
+            payload: {
+              currentSlide: 0,
+              transform: transformMap[0],
+            },
           })
         }
 
@@ -163,15 +166,15 @@ const SliderTrack: FC<Props> = ({
           dispatch({
             type: 'ADJUST_CURRENT_SLIDE',
             payload: {
-              currentSlide: totalItems - slidesPerPage,
-              transform: transformMap[totalItems - slidesPerPage],
+              currentSlide: currentSlide + totalItems,
+              transform: transformMap[currentSlide + totalItems],
             },
           })
           groupDispatch?.({
             type: 'SLIDE',
             payload: {
-              currentSlide: totalItems - slidesPerPage,
-              transform: transformMap[totalItems - slidesPerPage],
+              currentSlide: currentSlide + totalItems,
+              transform: transformMap[currentSlide + totalItems],
             },
           })
         }